Removed couchDB, importer, setup cli, and logins
Removed all references to couchDB throughout the repository. This
includes removing couchDB and couchapp from the package.json file, most
of the js files and the whole db dir under /server. In it's place, dummy
data was added to the api.js file under /server/routes to fake a DB
backend but always returns the same data. This will be replaces with the
new epochtalk-api repo at a later date.

Removed the importer because it will exist i the epochtalk-api repo.

Stubbed out CLI setup tool for now.

Updated the passport.js file to always return a valid user.

added bundle.min.js to the .gitignore file.