Added the Editor
Added all the editor related files (CSS/Templates/JS) to the appropriate
directories. As well as adding the medium-editor node package. Also
added a bash script to copy the css files over from the medium-editor
package to the app/css dir. This script is also added to the install
script in the package.json file.