Switch to Hapi
Switched from Express to Hapi. Also added all the routes from core
through a require().

The work invovled in switching to Hapi, also required moving some of the
routes around. All static assets are now found on /static and all core
routes are found in /api. This also required updating the gitignore and
the package.json file so that browserify builds to the right directory.
The index.html file was also updated to load the js file from the
correct dir.

server/index.js was updated to require, configure, and run Hapi now. A
few hapi security configs were also set. As for the routes, all the old
routes were deleted and only a few local routes remain. As mentioned
before, the /static and /api routes were set up. Any other route is
automatically sent to the index.html file for angular to handle.