User login and logout
Newly installed modules:
hapi-auth-jsonwebtoken
jsonwebtoken
redis and hiredis

Changed the configs to include the default redis host and post, also
added a privateKey property to use for jwt.

Added two new routes, /api/login and /api/logout to log users in using
their username and password and handling them back a jwt token to use
from then on. This token is saved in redis as our server side session
store. On logout, the token is deleted from redis. This is out way of
logging a user's session out.