Client-side login/logout
Added the client-side login/logout code. This involved creating a
authInterceptor to intercept all calls back to the server and injecting
the Authorization header with the appropriate token. This token is
created when the user logs in and is saved to either session or local
storage based on whether they click on the remember me checkbox.

The HTML inputs to handle the username/password/remember me were added
to the index.html page and the css was edited to make the input look
reasonably okay.

The Auth service and User resources were also edited to match the
implementation described above. Currently, login and logout work well
but registration is not coded up yet. The Auth Service is what handles
saving the token and username to storage upon successful login with the
hapi backend. It also handles removing those pieces of information on
logout. There are two other functions, isAuthenticated and currentUser
that return whether the user is authenticated by checking a token exists
in either storage, and the username for the currently signed in user.
The User resource currently holds the url for login, logout, and
registration and returns the content of those calls.