Fixed flickering screen issue on page change
Fixed an issue with the screen flickering during page changes. The issue
comes from the fact that the footer would jump to the top of the page
when there is no content during page changes. A fixed min-height has
been added to the ng-view div element to keep the footer outside the
range of the page.

Also removed the extra margin at the bottom of the footer.