User Registration
Added a user registration page at /register. This required creating a
registration controller and html page and changing the app/app.js file
and config.js files to point to the correct files.

The Auth service was also edited to have a registration method that
calls User resource which ends up calling the hapi service with the user
information to create a user.

There needs to be more validation in the registration controller since
right now it just allows user to create any username or use any email.