User Views
Added user views to the frontend. This commit is linked to the commit in
the core repo. Basically this commit adds the proper code to track a
logged in user's view of a thread. The datetime of the thread view is
logged in the user_views db metadata (threadId - timestamp). Each user
has their own user_views object that tracks the complete list of threads
they have viewed and at which time.

When a user views a list of threads via the /boards/{thread_id} path,
the complete list of user views is retrieve and compared against the
threads returned. The threads that have a new post later than when the
user last viewed this thread, has a new property called has_new_post
appended to the thread object and returned to the view.

In the view, if the has_new_post property is found, another property
called title_class is appended to the thread object with the value of
bold-title. This is a CSS class that will bold the title of the thread
if there is a new post that the user has not seen.