isAuthenticated
Added a method on both the client-side and server-side that allows the
client to check in with ther server to see if it is still logged in. The
server provides a route at '/api/authenticated' that the client queries
and if the Authorization header comes back as a Bearer token, then
authentication continues, if it comes back as 'Revoked' then the UI
reflects that change.