Posting from forum
Made the necessary changes to allow the editor to finally post from the
client-side to the server side. This required hooking up the editor
directive to the posts controller so that the posts controller can
retrieve the written information from the user and package it into a
JSON version of a post. This is then sent to ther server to be saved to
the database.

The User's id is grabbed from the authentication within Hapi while the
rest of the information is provided by angular. (title, body,
thread_id). For now, only the processed version of the body is sent to
the server until the database can handle storing both the processed and
encoded body text.

Aesthetic changes were also made to the editor to better align with the
look and feel of what we already have.