New Thread functionality
Added the ability to create new threads in a board. The first thing
added was a link to the new thread page from the /baord view. This
allows the user to create a thread from any board view.

A new view was added that takes in all the information needed to create
a new thread and the first post of that thread. This view is almost the
same as the new post view found in posts. The difference is that the
board id is also needed to create the thread first.

On server side, the POST /threads route was updated to take in the new
information, and creates a thread first, then using the newly created
thread_id, creates the first post for that thread. The new post is sent
back to the user.

Back on client side, once the POST call returns, the user is redirected
to the post listing view for the newly created thread.