Edit user posts
Added the ability to edit all posts that a user has previously created.
This new functionality depends on an update on core that returns the
user id for each post. This is the key that is used to determine whether
a user can edit that post or not.

From there, the user will now see an edit button next to all posts that
belong to them. Clicking that button brings up an in place edit of the
post using the epoch editor. The cancel button will return the post to
its previous state while the save button will update the post on server
side and trigger a reload of the page.

All error output are present as well. Including guards that will prevent
users from editing other user's posts.