Fix for userviews for a new user
There was a bug where a new user viewing a list of thread for a board
would hang the server. This is because the server would look for the
userviews object for that user but it doesn't exist yet. It now catches
that exception and returns an empty object until that userviews object
is created.