Bug Fix: edited post content not showing up
Fixed a bug introduced in the last commit where the previously type
input would not show up when the edit button was clicked. This only
happened on posts where the content did not contain BBCode. The problem
was the body was not being properly copied over to the encodedBody
property on scope when the edit button was clicked.