Profle Editing
Added the ability for users to edit their profile information from their
profile page. Their profile page is reached through the
/profiles/:username url. Currently the user can only modify their email.
There is the capacity to edit their username and password, to be added
later. Also, the ability to edit their other information will be added
once core can support it.

This feature required re-writing the profile path in both angular (with
the use of resolvers) and in hapi (fully written from the path down to
the config, auth, and validation schema).