Removed greeting messages, update username
Removed all code related to have a greeting message when a user logs in
since this is no longer being used anymore.

Also bound a setUser function to the Auth service that allows changing
the username being displayed in the header. This was added so that when
a user changes their username from a profile update, the new username is
reflected in the header.