Unique Email Validation directive
Created a directive that validates an email input field. This directive
checks that the email is not empty, is not the same as the original
email, and is unique according to the backend. It accomplishes the last
check by calling the backend which ends up querying to see if a user
with that email exists.