Updated Angular Styling for boards/main/register
For boards/main/register, the controller, template, and the config file
was updated to follow the lastest angular style guide. This requires
the use of the controllerAs syntax and the resolver pattern to
modularize the controller and to load it with the required params before
the controller is fired off. The template files have all been updated to
use ng-bind where necessary to control page flickering.