Angular Style updates to board, newThread, and posts, also fixes
Updated the above controllers, templates and the config file so they are
inline with the most up to date Angular Style guides. This includes
using controllerAs syntax, resolvers, reloadOnSearch for pagination, and
to load the params into the controller rather than fetching them after
the controller and view are loaded. The templates all utilize ng-bind as
well.

The pagination directive needed to be updated to allow for the updated
styles since the data may not be readily available by the time it
reached the pagination directive. It has now been changed to use two way
binded attrs params and reloaded the pages as those values change.