Clean up styles, updated HeaderCtrl
Updated the Header Ctrl to match the latest Angular styles. This
included the controllerAs syntax and ng-binding. The profiles page was
also updated to use ng-bind where possible.

Other files were cleaned up, all extraneous or otherwise un-used params
were removed.

A few fixes were added for errors in previous commits: pagination did
need a ng-bind-html in it and categories in boards do not have anything
for track by to latch on to.