Updated Editor Directive + posting validation
Updated the editor directive to two way bind the post body and encoded
body to the model the editor is attached to. This also handles the
business logic that dictates that the encodedBody is only used to
display the old text to the user to work off of. After the user edits
the post, the user text is saved to the body model. This effectively
hides all that ugliness from the controller.

All controller code that used to handle this has also been removed.

All the editors in posts and threads now have validation that disallows
the user to post unless both the title and the body is filled. It also
checks that the user is logged in for the newThread page.

The CSS for all forms have been condensed into one set of CSS.

The validation rules for the threads and post APIs have been updated to
account for post.encodedBody that is now an empty string.