Reset Editor + post page optimization
Updated the editor with the ability to reset the editor if need be. This
required adding an extra scope var to set $watch on, called, reset. If
this value is ever true, then the editor will re-initialize with
whatever is bound to the body and the encodedBody scope vars. This
allows to reset to a blank editor or some pre-defined text based on
what the model has been set to prior to calling reset.

This change has been implemented in both the posts and newThread
templates. The new post editor in the posts template is the only editor
currently making use of the reset switch. The switch allows for
optimization in the posts controller. Instead of reloading the page
whenever a new post is made or if a post is edited, only the API to
create or edit the post is made. An additional call to get all posts for
a specific page is also made due to the fact that a new post could
create a new page.