Updated the editor to handle foreign langs + walls of text
Added a few extra steps to the input parsing process. It now escapes all
special characters for security but the & character is added back in to
support unicode characters. Then each character is checked to see if it
higher than 127 in the character encoding, if it is, then the character
is converted into it's decimal form for saving. Finally the BBCode
parser is run to convert all BBcode into html. This parsed text is then
shown to the user. The raw text (everything minus the BBCode parsing) is
what's saved to scope which is eventually saved to the backend.

Also, a debounce was added to the onChange binding in the editor. This
was added because for large amounts of text, the constant parsing
overloaded the browser and could cause it to crash. The debounce offsets
this by only initiating the parser 250 milliseconds after the last
keypress. Any keypress within the 250 millisecond period just resets the
debounce for another 250 milliseconds.