Fix for safari's lack of browser storage
On the newer safari browsers, both local and session storage has been
disabled in private browsing mode. This commit creates a internal
'privateStorage' object on window that will act as a transient storage
to take place of session storage during that time.

It has been added to all the places in Auth and AuthInterceptor to make
the site work properly in private browsing mode for safari.

Also, the get/set User methods have been renamed to get/set Username to
be more appropriate to what the method actually does.

closes #13