Image Cdn server
Added an image cdn server that is run along with the main epochtalk
server. This image cdn server basically serves to act as a proxy to
route hotlinked images from another domain. This adds a layer of
security by putting all hotlinked images behind another domain. The
hotlink image urls are generated by the posts pre-processor. The url
scheme for images turns the generated urls into a url for the server to
retrieve the images from and pass on to the user.

The url scheme is as such /images/{method hash}/{url cipher}
The method hash indicates if this image is a hotlinked image. It is
basically just the word 'hotlink' with a sha256 hash.
The url cipher is the url of the image encrypted with AES256. This was
chosen because it can then be decrypted on the image cdn and the url can
be routed through the image server.