Expanded the editor to handle file uplaods
The Editor has been rewritten to allow more content and tools to fix
into the directive. The first step was to update the CSS to utilize
foundation. The second step was to updated the HTML and pull in the
editor and preview titles into the editor rather than have them defined
outside the editor every time.

One service and one extra directive was created to support file uploads.
The directive allows access to HTML5's File API to keep track of files
selected by the user from their hard drive. The service is a generic
upload service that uses FormData to upload the file to the image cdn
server. This may be removed later so that uploads go directly to s3.