Editor: handle drag and drop uploads
Updated the editor to handle drag and drop uploads. This required making
the entire editor footer a drag and drop zone. The required event
listeners have been added to handle dragEnter, dragLeave, and drop. The
drop listener pulls out the file list from the drop event and calls the
upload function with an array of images.