Editor: handle images only
Updated the editor to handle only image uploads. The input[type=file]
was updated with an additional attribute 'accept="image/*"' and will
only accept images types now. The drag and drop implementation checks
the file type of each incoming file and only uploads files that are of
an image type.