Added Images uploaded to editor footer
Added a section to the footer that shows the user the files being
uploaded and when the upload is done, the display updates with a link
to the uploaded file.

This commit also fixes a problem where the same file input's filelist is
not cleared after uploading. It is now cleared after each upload.