Edit Post: replace post with new content from server
This commit addresses an issue where after a post with an image is
edited, the body of the post is left with the original hotlinked image
src rather than the server generated one. This commit fixes that by
replacing the body and encodedBody with the values returned by the
server.