Pulled Editor into a fixed position modal
This commit encapsulates all the work needed to pull the editor
directive out into a fixed position modal. This includes updating the
scss to contain the editor in a fixed position div as well as updating
the styles to fit correctly.

The controller was also updated to handle new and previous post editing
into a single modal rather than having an editor for all posts on a
page. This change also updated the controller so that there is only one
method to set the properties of an editor and one method to save any
changes to a new or old post.

The posts.data.html was also updated to account for the single editor.