Bug Fix: posts reload after new post
This fixes an issue where after a new post is made, the page wouldn't
refresh with new content. This was fixed by retrieving the last page of
posts, updating the current page in the pagination module, and scrolling
to the new post. This is all done within the posts controller.