Editor: pulled posts editor into parent controller
Pulled the post editor into the parent controller so that it maintains
state across all the pages. This required re-writing the child
controller, 'posts.controller.js' and creating a new parent controller,
'parent.controller.js'

The style and html were also updated to give the toolbar at the bottom a
cleaner look and feel.