Scrolling Bug fix pulled into directive
Pulled the scrolling bug fix into a directive, 'scroll-lock'. This
directive is an attribute only directive and is added to any element
that can scroll. Once added, it'll add the body-scroll-lock class to the
body element while the mouse is in the element. This will prevent the
body from accepting the scroll event bubble up. When the mouse leaves
the element the css class is removed. The class is also removed on
element destruction.