Editor: Dynamically resize editor
Added a feature where the top of the editor overlay is now a 'grab-able'
handle that can be used by the user to dynamically resize the size of
the editor overlay. This required creating a new directive that watches
the mousedown, mouseup, and mousemove events, to determine whether to
set the cursor to 'ns-resize' and set a new height for the edtior
overlay. It also uses requestAnimationFrame to do all the drawing thus
keeping the animations at 60 fps.

The majority of the events basically watch the mouse position to see if
it near the top over of the overlay and set the appropriate cursor icon.
If the user clicks on the element while the in the sweet spot of the
element, the directive calculates the y distance of any further mouse
movement and determines the appropriate height for the editor. It then
sets the new height variable and lets requestAnimationFrame pick that
variable up when it's ready and repaints the page.