Editor Bug: focus is never set
Fixed a bug in the editor where the focus flag doesn't actually set the
focus on the editor. Apparently contenteditable divs do not react the
same way regular inputs do with the .focus() method. Instead, grabbing
the $window.selection, collapsing then adding the range, then calling
the .focus() method is what needs to be done.