Editor optimization: set focus on editor show
Created a dedicated focus flag that is bound to the editor's focus
attribute. This flag is reset by the editor after the focus has been
set. The flag itself is set to true whenever the edtior is opened or
reset by the the edit button or the new post button.

The function to reset the editor is now called, 'initEditor' and a new
function was created to open the editor and focus the input called,
'openEditor.'