Editor: only editor inputs scrolls
Refactored the editor container on the posts page so that the only
element with the ability to scroll is the input div. This allows all the
other content in the editor and it's parent container to stay in view.
Much of the CSS had to be rewritten with hard values to account for
varying height of the input div. Also, the hard limit on the max height
of the editor container has been removed.