Fix for new thread submission
Created a fix to the issue where on submitting a new thread, the editor
exiting alert is fired. A new optional attribute was added to the editor
so that the controller can notify the directive that the user is now
submitting the editor content and is being redirected. This notification
turns off the editor exiting alert on the editor side.