Convert quote date param on post import
A new pre-processing method was created that will convert all date
params found in quote BBCode tags in the raw_body of posts. This method
is run before the parsing of the BBCode tags. This method converts the
date params from the second scale to the millisecond scale because JS's
Date object expects numbers in the millisecond scale rather than the
second scale.