Bug Fix: Page exit function fired with no editor change
Fix a bug where the page exit function would fire and send the user an
alert even if there was no change in the editor text. The reason this
bug occurred is because on init, there is a condition where scope.body
is the only text var populated. The page exit function checks whether
scope.rawBody has been changed, thus is appears to the page exit
function that the text has changed even if it hasn't.

In essence, a var scope.originalText is kept with the initial text
value. The page exit function will check the rawBody against this var.
But in certain cases (when only the scope.body has a value),
scope.rawBody is undefined. Thus scope.originalText has the initial text
value while scope.rawBody has a value of undefined thus triggering the
alert.

This has been fixed by populating scope.rawBody in the case that
scope.body is the only text var that has a value.