Bug Fix: Editor allowed exit dispite alert
The editor allowed a user to leave when the user clicked cancel on the
redirect alert. The expected case was to keep the user on the current
page. This occurred because the stateChangeStart event only has the
preventDefault method but the stopPropagation method was also called
causing an exception and allowing the user to leave.

This has been fixed by only calling the preventDefault method on the
event.