Dirty Flag on editor, alerts on loading new post
The alert that usually stops the user from leaving the page when there's
new content in the editor has also been applied to all the buttons that
load another or new post into the editor. This is to ensure that the
user cannot overwrite anything they've written but have not saved yet.

In the course of writing that functionality, a dirty flag scope variable
was created to easily signify if the content in the editor has been
touched. This dirty flag is now used in other parts of the page such as
the send button validation.