Scroll-Lock: Scroll fix/hijack
The scroll-lock directive needed to be re-written to account for mac's
setting that allows the scroll bars to always be present. In this
scenario, the body's overflow: hidden trick would not work because it
would cause the scroll bars to flash on and off when the user's mouse
would hover over the editor and back to the posts.

To fix this problem, the scroll-lock directive now hijacks the
mousewheel event and sets the scroll position on its own. It then calls
preventDefault and stopPropagation to keep the body from scrolling.
Since the directive lives directly on the element being locked, the attr
to turn this on and off is no longer needed as well.