Editor: image listing modal
Updated how user uploaded images are shown to the user. Originally, they
were just listed but with how the editor is setup, this list could
expand down past the viewable area. This area could not be scrolled due
to the containing div's positioning. So a modal is used instead to
display all the images when the number of images is larger than one. If
there's only one image, then that link will be shown.