Hid unused post action buttons, better logic for those used
Hid all the extra post action buttons that are not being used and added
better logic to determine when certain buttons should be shown. At this
point, all buttons are hidden until a user logs in. Upon which the quote
button will be shown. If a user has created a post, the edit button will
also be shown.