Auto-link and Auto-date added to editor preview, posts
Added the auto-link and auto-date directives to the editor preview so
that the preview output will exactly match what the contents will look
like when posted. It has also been added to the post listing to update
quote dates with human readable dates.

The directives being used in the editor preview are also bound to scope
variables that are essentially boolean flags. This allows the editor to
rerun the auto-* directives at the end of each parse "cycle."