BBCode style stitching directive
A new directive that applies the bbcode styles using ng-style and style
hints from the span classes. This directive sits on top of each post and
watches the post's html. On change, it will search through the html
looking for any element that matches the pattern 'class="bbcode-*-*"'
This pattern is then replaced with ng-style='*' where the * represents a
new element style that is generated from the element's class. The class
attribute itself is replace completely.

After the element's html is replaced. $compile is called so that angular
can pick up the ng-style attributes and safely create the new styles for
this element.

This was mainly created because any style tags added on by the bbcode
parser are stripped by angular and need to be inserted manually again.
Inserting the the style through ng-style allows angular to securely
evaluate the styles before actually added them to element.

The auto-link and the auto-date directives have also been updated to
watch the element's text for changes rather than relying on a boolean
flag. Therefore, the editor was also updated to have the auto-* flags
removed since they are no longer needed.