Updated Profile page
Updated the look and feel of the profile page. Avatars don't have any
validations as of this moment and all profile field editing is done in a
single modal otherwise. Signatures should be broken out to a separate
modal when possible. Also, chart and chart data is hard coded till
actual user data is available.