Fix for quoting issue, #51
This fixes an issue with the quote text being shifted outside the quote
div when there is a new paragraph before the quote BBCode. This is
mainly an issue with the Medium error so the Medium editor was removed
until it can be implemented fully in it's own editor later.

Closes #51