Added LoginRequired switch
Added a switch in the config that turns the forum private or public.
Private meaning that the user needs to be signed in to view any content
on the site. This required added a pre method, requireLogin, to each
'GET' type route that checks whether the switch is set and whether the
user is logged in. It returns that boolean to the handler method which
then best decides how to return to the requester (array or empty
object).