Bug fix: Disply 0 posts if none found
Before this commit, the profile page for a user would show 'Posts' if
the user hasn't made any posts yet. This has been fixed so that if a
post_count isn't returned for that user, it will instead display '0
Posts.' This was previously fixed but was missed during the commit that
actually implemented post_count.