Fixes #21, #22
This fixes all the issues with ui-router in relation to handling the
hash in the url, the loading of posts across pages, and going to the
right page after making a new post. The main issue is that
reloadOnSearch was not implemented until ui-router 0.2.13. After
updating, reloadOnSearch was set on the posts states which does not
reload the controller on changes to the query parameters or hash.

This required firing an event on locationChangeSuccess from ui-router.
In this event, the page and limit vars are pulled down and the correct
posts are loaded on to the page using the pullPost method. The hash in
the url is also handled using anchorScroll now instead of uiViewScroll.

Also, the index state was deleted since it's no longer used.

closes #21
closes #22