Optimized posts controllers
There was a few unused variables, orphaned event handlers, and request
optimizations that could be done around the controllers that lead to
less calls to ther server overall. The main focus was to ensure that the
locationChangeSuccess event handler was removed on controller
destruction. The other item of interest is that an extra call is being
made to get the thread on each page change. This is to ensure that the
post page count is accurate at all times.