Board controller paging updated to match posts
The Board controller was updated to match how the posts controllers are
handling page updates. Instead of reloading the controller on each page
change, only the threads for that page and the board object is being
requested. This cuts down on the resources that need to be pulled down
on each page change.