Removed image server - optimized images code
Removed the image server dependency and refactored the images code so
that the image url generated for the user is now just retrieved from the
server configs. This implies that the image hosting does not need to be
proxied through another server but rather the user can retrieve the
images directly from the image host.

The major breaking changes are mainly in the config file which has been
refactored to have a image sub object. The image code in general has
been consolidated into one file in server/images.js. The images
themselves now retain their file extensions as well.

The procfiles and the package.json files have been updated to remove the
image server dependecy.