Better image handling, improvements to posts/editing
This started out as a way to handle images that were being uploaded by
the server when a hotlink is provided in a new post. The idea was to
retry the image until the server was done uploading. The issue with this
is that adding a new directive to the post was near impossible with how
the other post directives were setup.

Thus the auto-link, auto-date, and the style-fix directives were
collapsed into one directive that did all the post preprocessing in one
loop rather than each in their own event loops. The image reloading was
pulled into it's own directive that lives on each image tag itself and
handles retrying images if they fail.

The editor was also updated to use the new post-processing directive
with a switch to disable style-fixing as it's not needed. Also the
editor's parsing was updated. It no longer injects the parsed html into
the preview element but rather let's angular updated the post-processing
directive which goes on to updated the preview element's html.

Updates were also made to the images.js file to allow GIFs to be
uploaded.

A loading gif was also added to the public/img dir (temporary)