Bug Fix: loginRequired should handle false for a env var
The loginRequired config variable could not handle 'false' as an env var
because it sees that value as a string and thus it evaluates to true.
It's been fixed by putting the env var through a function that checks
whether the string matches any of the following: true, True, TRUE, 1.

The method that actually enforces the loginRequired has been moved out
to a server method on hapi to cut down on extraneous code. All the
config method and pre files have been updated as per this change.

Path.join has been updated to Path.normalize where files were touched.