Fix for bbcode in signature
This is a temporary fix to issue #57. The reason why the color wasn't
showing up was due to angular not re-rendering the signature dom element
with the new post-processing & style fix directives.

Also, images have been removed from signatures on the backend after all
the signature processing is complete.

This work does bring up an issue with editing signatures. Since the
BBCode editor is not being used at this point and the user's input is
not being saved on the backend, repeated editing of the signature is not
a plensant experience.

closes #57