S3 Bucket creation if it doesn't exist
Added a check on startup that will check if the S3 bucket exists.
Pkgcloud was removed in favor of using the aws-sdk instead because
pkgcloud was not granular enough. Extra check were also added to the
server side image upload to ensure that the content-type and
content-length are valid as per the configs.

Also fixed an issue with deleting files after image expiration.